Use curl in pkgget
curl is available on MacOS as well as on Linux, and it is already used in
IRAF. And its sources are currently in IRAF.

In contrast, pkgget used `/usr/bin/ftp` on MacOS, which was removed in 10.13
(High Sierra). Instead of distributing a binary "ftp" executable for those
systems, this patch replaces it with curl on all systems.