Limit number of finfo structs returned by KI_ZFINFO to MAX_ARGS
p_arg is defined to have a length of MAX_ARGS (=13), and this is the
maximum number of finfo structs that can be returned; even when zfinfo
returns LEN_FINFO values. To be flexible here when MAX_ARGS or
LEN_FINFO is changed, we use the min() of these two.